Рефактор RoleService

+++ src/main/java/ru/innopolis/stc9/susev/services/RoleService.java
@@ -1,5 +1,6 @@
 package ru.innopolis.stc9.susev.services;
 
+import com.sun.istack.internal.Nullable;
 import ru.innopolis.stc9.susev.db.dao.IRoleDao;
 import ru.innopolis.stc9.susev.db.dao.RoleDao;
 import ru.innopolis.stc9.susev.pojo.Role;
@@ -7,33 +8,45 @@ import ru.innopolis.stc9.susev.pojo.Role;
 import java.sql.SQLException;
 import java.util.List;
 
-public class RoleService {
+public class RoleService implements IRoleService {
     private IRoleDao roleDao = new RoleDao();
 
-    public List<Role> getRoles(){
+    @Override
+    @Nullable
+    public List<Role> getRoles() {
         try {
             return roleDao.getRoles();
         } catch (SQLException e) {
-            // TODO: 12.05.2018 обработка
+            logger.warn("Ошибка получения всех ролей", e);
+            return null;
         }
-        return null;
     }
 
-    public boolean createRole(Role role){
+    @Override
+    @Nullable
+    public boolean createRole(Role role) {
+        if (role == null) {
+            return false;
+        }
         try {
             return roleDao.addRole(role);
         } catch (SQLException e) {
-            // TODO: 12.05.2018 обработка
+            logger.warn("Ошибка создания роли", e);
+            return false;
         }
-        return false;
     }
 
-    public boolean deleteRole(int id){
+    @Override
+    @Nullable
+    public boolean deleteRole(int id) {
+        if (id == 0) {
+            return false;
+        }
         try {
             return roleDao.deleteRole(id);
         } catch (SQLException e) {
-            // TODO: 12.05.2018 обработка
+            logger.warn("Ошибка удаления роли", e);
+            return false;
         }
-        return false;
     }
 }
